requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
anthropic>=0.18.0
python-dateutil>=2.8.0
